Rename `generator_poly_to_matrix()` to `poly_to_generator_matrix()`
