+++ galois/code/__init__.py
@@ -2,5 +2,5 @@
 A subpackage containing forward error correction codes.
 """
 from .bch import *
-from .common import *
+from .cyclic import *
 from .reed_solomon import *
